requirements.txt
python-binance~=1.0.10
asyncio
pandas
tqdm
feedparser